tide closer with traits.h
