+++ shared/public/system.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "config.h"
-#include "traits.h"
 
 #ifdef INK_ENABLE_UNREAL
 #include "Misc/AssertionMacros.h"
@@ -13,7 +12,6 @@
 #include <exception>
 #include <stdexcept>
 #include <optional>
-#include <tuple>
 #endif
 
 namespace ink
@@ -119,138 +117,13 @@ namespace ink
 	};
 #endif
 
-#ifdef INK_ENABLE_STL
-	template<typename ... Tys>
-	using tuple = std::tuple<Tys...>;
-	template<typename T>
-	constexpr auto t_get = std::get<T>;
-#else
-	namespace tuple_internal {
-		template<size_t I, typename T>
-		class tuple_leaf{
-		public:
-			tuple_leaf() : _value() {};
-			template<typename U>
-			explicit tuple_leaf(U&& u) : _value(std::forward<U>(u)) {}
-			T& get() { return _value; }
-			const T& get() const { return _value; }
-		private:
-			T _value;
-			tuple_leaf(const tuple_leaf& tl) = delete;
-			tuple_leaf& operator=(const tuple_leaf&) = delete;
-		};
-
-		// handle indexing
-		template<size_t... Is>
-		struct tuple_indexes {};
-
-		template<size_t End, size_t Start = 0, size_t... Is>
-		struct make_tuple_indexes {
-			using type = typename make_tuple_indexes<End, Start+1, Is..., Start>::type;
-		};
-		template<size_t End, size_t... Is>
-		struct make_tuple_indexes<End, End, Is...> {
-			using type = tuple_indexes<Is...>;
-		};
-
-		// handle types
-		template<size_t I, typename T, typename ...Tys>
-		struct tuple_type : tuple_type<I-1,Tys...> {};
-		template<typename T, typename ...Tys>
-		struct tuple_type<0, T, Tys...> {
-			using type = T;
-		};
-
-		template<typename T, size_t I, typename U, typename ...Tys>
-		struct type_index_imp : type_index_imp<T,I+1,Tys...> {};
-		template<typename T, size_t I, typename ...Tys>
-		struct type_index_imp<T,I,T,Tys...> {
-			static constexpr size_t value = I;
-		};
-		template<typename T, typename ...Tys>
-		constexpr size_t type_index = type_index_imp<T,0,Tys...>::value;
-
-
-		template<typename Indexes, typename ...Tys>
-		struct tuple_imp;
-
-		template<size_t ...Is,  typename ...Tys>
-		struct tuple_imp<tuple_indexes<Is...>, Tys...>
-			: public tuple_leaf<Is, Tys>...
-		{
-			template<typename ...Us>
-			tuple_imp(Us&& ... us) : tuple_leaf<Is,Tys>(std::forward<Us>(us))... {
-				static_assert(sizeof...(Us) == sizeof...(Tys),
-						"Tuple must be initialized with same amount of arguments"
-						", then types!");
-			}
+	namespace runtime::internal {
+		template<typename T>
+		struct always_false {
+			static constexpr bool value = false;
 		};
 	}
 
-	/// minimal tuple class, only for simple data types!
-	template<typename ...Tys>
-	class tuple
-		: public tuple_internal::tuple_imp<
-		  	typename tuple_internal::make_tuple_indexes<sizeof...(Tys)>::type,
-		  	Tys... >
-	{
-		using base = tuple_internal::tuple_imp<
-			typename tuple_internal::make_tuple_indexes<sizeof...(Tys)>::type,
-			Tys...>;
-		using this_type = tuple<Tys...>;
-		template<size_t I>
-		using element_type = typename tuple_internal::tuple_type<I, Tys...>::type;
-		template<typename T>
-		static constexpr size_t type_index = tuple_internal::type_index<T, Tys...>;
-	public:
-		template<typename ...Us>
-		tuple(Us&& ... us) : base(std::forward<Us>(us)...) {}
-
-		template<size_t I>
-		friend
-		inline constexpr element_type<I> const&
-		t_get(const this_type& t);
-		template<typename T>
-		friend
-		const T&
-		t_get(const this_type& t);
-	};
-
-	template<size_t I, typename T>
-	inline constexpr typename T::template element_type<I> const&
-	t_get(const T& t) {
-		return static_cast<tuple_internal::tuple_leaf<I, typename T::template element_type<I>>const&>(t).get();
-	};
-
-	template<typename T, typename Tuple>
-	const T&
-	t_get(const Tuple& t) {
-		return t_get<Tuple::template type_index<T>>(t);
-	};
-
-	template<>
-	class tuple<> {
-	public:
-		tuple() {}
-	};
-#endif
-
-	namespace runtime::internal
-	{
-		struct false_type { static constexpr bool value = false; };
-		struct true_type { static constexpr bool value = true; };
-		template<typename T>
-		struct always_false : false_type {};
-
-		template<typename T, typename Tuple>
-		struct has_type;
-		template<typename T>
-		struct has_type<T,tuple<>> : false_type {};
-		template<typename T, typename U, typename ... Tys>
-		struct has_type<T, tuple<U, Tys...>> : has_type<T,tuple<Tys...>> {};
-		template<typename T, typename ... Tys>
-		struct has_type<T, tuple<T, Tys...>> : true_type {};
-	}
 
 #ifdef INK_ENABLE_STL
 	template<typename T>
